requirements.txt
gradio
moviepy
openai-whisper
torch